pyproject.toml
[tool.poetry]
name = "image-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
python-dotenv = "^0.19.2"
python-multipart = "^0.0.5"
uvicorn = "^0.17.6"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
